requirements.txt
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
motor==2.4.0
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.1
python-decouple==3.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.14.0
